refactor: move energy from application to domain
