+++ src/libecalc/domain/infrastructure/energy/component_energy_context.py
