+++ src/libecalc/domain/infrastructure/energy/emitter.py
@@ -1,9 +1,9 @@
 import abc
 from typing import Optional
 
-from libecalc.application.energy.component_energy_context import ComponentEnergyContext
-from libecalc.application.energy.energy_model import EnergyModel
 from libecalc.core.result.emission import EmissionResult
+from libecalc.domain.infrastructure.energy.component_energy_context import ComponentEnergyContext
+from libecalc.domain.infrastructure.energy.energy_model import EnergyModel
 
 
 class Emitter(abc.ABC):
